Fixed the sync validation issue

+++ src/main/java/ck/panda/service/DepartmentServiceImpl.java
@@ -238,34 +238,35 @@ public class DepartmentServiceImpl implements DepartmentService {
         Errors errors = validator.rejectIfNullEntity("department", department);
         errors = validator.validateEntity(department, errors);
         config.setServer(1L);
-        List<Project> projectResponse =  projectService.findByDepartmentAndIsActive(department.getId(), true);
-        List<VmInstance> vmResponse  =  vmService.findByDepartment(department.getId());
-        List<Role> roleResponse = roleService.findByDepartment(department);
-        List<Volume> volumeResponse = volumeService.findByDepartment(department.getId());
-        List<User> userResponse = userService.findByDepartment(department.getId());
-        if (projectResponse.size() != 0  || vmResponse.size() != 0 || roleResponse.size()!= 0 || volumeResponse.size() != 0 ) {
-         errors.addGlobalError( "You have following resources for this department: <br><ul><li>Project : " + projectResponse.size() +
-                    "</li><li>Instance : " +vmResponse.size()+
-                    "</li><li>Role : " +roleResponse.size()+
-                    "</li><li>Volume : " +volumeResponse.size() +
-                    "</li><li>User : " + userResponse.size() +
-                    "</li></ul><br>Kindly delete associated resources and try again");
+        if(department.getSyncFlag()) {
+			List<Project> projectResponse = projectService.findByDepartmentAndIsActive(department.getId(), true);
+			List<VmInstance> vmResponse = vmService.findByDepartment(department.getId());
+			List<Role> roleResponse = roleService.findByDepartment(department);
+			List<Volume> volumeResponse = volumeService.findByDepartment(department.getId());
+			List<User> userResponse = userService.findByDepartment(department.getId());
+			if (projectResponse.size() != 0 || vmResponse.size() != 0 || roleResponse.size() != 0
+					|| volumeResponse.size() != 0) {
+				errors.addGlobalError("You have following resources for this department: <br><ul><li>Project : "
+						+ projectResponse.size() + "</li><li>Instance : " + vmResponse.size() + "</li><li>Role : "
+						+ roleResponse.size() + "</li><li>Volume : " + volumeResponse.size() + "</li><li>User : "
+						+ userResponse.size() + "</li></ul><br>Kindly delete associated resources and try again");
 
+			}
         }
         if (errors.hasErrors()) {
             throw new ApplicationException(errors);
-        }
-        else {
-
+        } else {
              department.setIsActive(false);
              department.setStatus(Department.Status.DELETED);
-             String departmentResponse = csAccountService.deleteAccount(department.getUuid(), "json");
-             JSONObject jobId = new JSONObject(departmentResponse).getJSONObject("deleteaccountresponse");
-             if (jobId.has("jobid")) {
-                 String jobResponse = csAccountService.accountJobResult(jobId.getString("jobid"), "json");
-                 JSONObject jobresults = new JSONObject(jobResponse).getJSONObject("queryasyncjobresultresponse");
-            }
-             LOGGER.debug("Department deleted successfully" + department.getUserName());
+             if(department.getSyncFlag()) {
+				String departmentResponse = csAccountService.deleteAccount(department.getUuid(), "json");
+				JSONObject jobId = new JSONObject(departmentResponse).getJSONObject("deleteaccountresponse");
+				if (jobId.has("jobid")) {
+					String jobResponse = csAccountService.accountJobResult(jobId.getString("jobid"), "json");
+					JSONObject jobresults = new JSONObject(jobResponse).getJSONObject("queryasyncjobresultresponse");
+				}
+				LOGGER.debug("Department deleted successfully" + department.getUserName());
+             }
         }
 
         return departmentRepo.save(department);
